+++ src/metatrain/utils/io.py
@@ -8,8 +8,7 @@ from urllib.request import urlretrieve
 import torch
 from metatomic.torch import check_atomistic_model, load_atomistic_model
 
-from ..utils.architectures import find_all_architectures
-from .architectures import import_architecture
+from .architectures import find_all_architectures, import_architecture
 
 
 def check_file_extension(
@@ -63,7 +62,11 @@ def is_exported_file(path: str) -> bool:
         return False
 
 
-def _hf_hub_download_url(url: str, hf_token: Optional[str] = None) -> str:
+def _hf_hub_download_url(
+    url: str,
+    hf_token: Optional[str] = None,
+    cache_dir: Optional[Union[str, Path]] = None,
+) -> str:
     """Wrapper around `hf_hub_download` allowing passing the URL directly.
 
     Function is in inverse of `hf_hub_url`
@@ -106,9 +109,10 @@ def _hf_hub_download_url(url: str, hf_token: Optional[str] = None) -> str:
         filename=filename,
         subfolder=subfolder,
         repo_type=None,
+        cache_dir=cache_dir,
         revision=revision,
-        endpoint=endpoint,
         token=hf_token,
+        endpoint=endpoint,
     )
 
 
@@ -119,6 +123,10 @@ def load_model(
 ) -> Any:
     """Load checkpoints and exported models from an URL or a local file for inference.
 
+    Remote models are downloaded to a local cache directory, with hashed filenames to
+    avoid common name collisions. The cache directory path is logged during the loading
+    process.
+
     If an exported model should be loaded and requires compiled extensions, their
     location should be passed using the ``extensions_directory`` parameter.
 
@@ -127,8 +135,8 @@ def load_model(
 
     .. note::
 
-        This function is intended to load models for inference in Python. For continue
-        training or finetuning use metatrain's command line interfaace
+        This function is intended to load models only for **inference** in Python. To
+        continue training or to finetune use metatrain's command line interface.
 
     :param path: local or remote path to a model. For supported URL schemes see
         :py:class:`urllib.request`
@@ -137,9 +145,6 @@ def load_model(
     :param hf_token: HuggingFace API token to download (private) models from HuggingFace
 
     :raises ValueError: if ``path`` is a YAML option file and no model
-    :raises ValueError: if no ``archietcture_name`` is found in the checkpoint
-    :raises ValueError: if the ``architecture_name`` is not found in the available
-        architectures
     """
 
     if Path(path).suffix in [".yaml", ".yml"]:
@@ -148,14 +153,15 @@ def load_model(
         )
 
     path = str(path)
+    url = urlparse(path)
 
-    # Download remote model
-    # TODO(@PicoCentauri): Introduce caching for remote models
-    if urlparse(path).scheme:
-        if hf_token is None:
-            path, _ = urlretrieve(path)
+    if url.scheme:
+        if url.netloc == "huggingface.co":
+            path = _hf_hub_download_url(url=url.geturl(), hf_token=hf_token)
         else:
-            path = _hf_hub_download_url(path, hf_token=hf_token)
+            # Avoid caching generic URLs due to lack of a model hash for proper cache
+            # invalidation
+            path, _ = urlretrieve(url=url.geturl())
 
     if is_exported_file(path):
         return load_atomistic_model(path, extensions_directory=extensions_directory)