+++ docs/src/dev-docs/changelog.rst
@@ -27,6 +27,8 @@ Unreleased
 Added
 #####
 
+- Added caching for ``load_model`` when downloading from URLs. Models will be stored in
+  a cache directory, which is logged when the model is loaded.
 - ``extra_data`` is now a valid section in the ``options.yaml`` file, allowing users to
   add custom data to the training set. The data is contained in the dataloader and can
   be used in custom loss functions or models.